KILL [CONNECTION | QUERY] support, part2B
Various small changes to part2, as suggested by comments and otherwise.
Mostly renaming, working logic should not change.

Exception: session id changed to 64bit in the container and associated
functions. Another commit will change it to 64bit in the session itself.

+++ include/maxscale/session.h
@@ -416,10 +416,11 @@ void session_clear_stmt(MXS_SESSION *session);
 /**
  * Try to kill a specific session. This function only sends messages to
  * worker threads without waiting for the result.
+ *
  * @param issuer The session where the command originates.
  * @param target_id Target session id.
  */
-void session_broadcast_kill_command(MXS_SESSION* issuer, uint32_t target_id);
+void session_broadcast_kill_command(MXS_SESSION* issuer, uint64_t target_id);
 
 /**
  * @brief Convert a session to JSON
